+++ config/passport.js
@@ -44,5 +44,5 @@ exports.isAuthenticated = (req, res, next) => {
   if (req.isAuthenticated()) {
     return next();
   }
-  res.redirect('http://login.inkdby.com');
+  res.redirect('http://login.example.com');
 };